2.0.3

+++ buildSrc/src/main/java/baritone/gradle/task/CreateDistTask.java
@@ -21,11 +21,8 @@ import org.gradle.api.tasks.TaskAction;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.nio.file.OpenOption;
 import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.security.MessageDigest;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -55,11 +52,10 @@ public class CreateDistTask extends BaritoneGradleTask {
             Files.createDirectory(dir);
         }
 
-        // Copy build jars to dist/
-        // TODO: dont copy files that dont exist
-        Files.copy(this.artifactApiPath, api, REPLACE_EXISTING);
-        Files.copy(this.artifactStandalonePath, standalone, REPLACE_EXISTING);
-        Files.copy(this.artifactUnoptimizedPath, unoptimized, REPLACE_EXISTING);
+        // Copy build jars to dist/ if they exist
+        if (Files.exists(this.artifactApiPath)) Files.copy(this.artifactApiPath, api, REPLACE_EXISTING);
+        if (Files.exists(this.artifactStandalonePath)) Files.copy(this.artifactStandalonePath, standalone, REPLACE_EXISTING);
+        if (Files.exists(this.artifactUnoptimizedPath)) Files.copy(this.artifactUnoptimizedPath, unoptimized, REPLACE_EXISTING);
 
         // Calculate all checksums and format them like "shasum"
         List<String> shasum = Files.list(getRootRelativeFile("dist/"))
